feat: aplicar tradução automática global por geolocalização
